feat: make location soil id card display selected soil id. refactor for clarity.
